Introduced Supervisor Processes
These take the place of the old Monitor module. Now supervisors
are a process like all the others. The weakest part of the
design is that kill by supervisor is done by using [Lwt.cancel].
If this were to be called in any other way other than by the
corresponding supervisor, the supervisor book-keeping will get
confused.