Eliminated Receiver
We are starting to migrate to a system without supervisors.
Receiver is now launched as an async thread in Peer. If it dies,
it posts a message to the peer (ReceiverAborted), so that the peer
can abort its own operation.

Also: we now send BitField messages after handshake (if we have at
least one piece).