UDP tracker now uses plain UDP sockets
Before it was using buffered Lwt_io.channels, and we were writing
directly into the channel, which makes me wonder if any UDP packet
got sent correctly.

To help with this shift to a 'lower' level appraoch, modules Get
and Put were written to parse and construct (resp.) strings received
and sent on the sockets. The code for binary IO was copied from
Ocplib-endian and only works in OCaml 4.01+. In fact, we might as well
use that library.